[REFACTOR] - Auth data layer 로 이동
