+++ app/src/main/java/com/jjsh/smartshopping/data/repository/CartItemRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package com.jjsh.smartshopping.data.repository
 
-import com.jjsh.smartshopping.common.Auth
+import com.jjsh.smartshopping.data.auth.Auth
 import com.jjsh.smartshopping.data.local.datasource.LocalDataSource
 import com.jjsh.smartshopping.data.local.dto.toDto
 import com.jjsh.smartshopping.domain.model.CartItem
